+++ cumulusci/newcli/cli.py
@@ -1,9 +1,13 @@
 import json
 import os
+import sys
 import webbrowser
+import code
 
 import click
+from plaintable import Table
 
+import cumulusci
 from cumulusci.core.config import YamlGlobalConfig
 from cumulusci.core.config import YamlProjectConfig
 from cumulusci.core.config import ConnectedAppOAuthConfig
@@ -17,7 +21,9 @@ from cumulusci.core.exceptions import GithubNotConfigured
 from cumulusci.core.exceptions import KeychainConnectedAppNotFound
 from cumulusci.core.exceptions import KeychainKeyNotFound
 from cumulusci.core.exceptions import MrbelvedereNotConfigured
+from cumulusci.core.exceptions import NotInProject
 from cumulusci.core.exceptions import ProjectConfigNotFound
+from cumulusci.core.exceptions import TaskOptionsError
 from cumulusci.core.exceptions import TaskRequiresSalesforceOrg
 from cumulusci.core.utils import import_class
 
@@ -37,15 +43,25 @@ class CliConfig(object):
         self._load_global_config()
         self._load_project_config()
         self._load_keychain()
+        self._add_repo_to_path()
+
+    def _add_repo_to_path(self):
+        if self.project_config:
+            sys.path.append(self.project_config.repo_root)
 
     def _load_global_config(self):
-        self.global_config = YamlGlobalConfig()
+        try:
+            self.global_config = YamlGlobalConfig()
+        except NotInProject as e:
+            raise click.UsageError(e.message)
 
     def _load_project_config(self):
         try:
             self.project_config = self.global_config.get_project_config()
         except ProjectConfigNotFound:
             pass
+        except NotInProject as e:
+            raise click.UsageError(e.message)
 
     def _load_keychain(self):
         self.keychain_key = os.environ.get('CUMULUSCI_KEY')
@@ -57,13 +73,20 @@ class CliConfig(object):
             self.keychain = self.keychain_class(self.project_config, self.keychain_key)
             self.project_config.set_keychain(self.keychain)
 
+try:
+    CLI_CONFIG = CliConfig()
+except click.UsageError as e:
+    click.echo(e.message)
+    sys.exit(1)
+
+
 pass_config = click.make_pass_decorator(CliConfig, ensure=True)
 
 def check_connected_app(config):
     check_keychain(config)
     if not config.keychain.get_connected_app():
         raise click.UsageError("Please use the 'org config_connected_app' command to configure the OAuth Connected App to use for this project's keychain")
-        
+
 
 def check_keychain(config):
     check_project_config(config)
@@ -80,7 +103,16 @@ def check_project_config(config):
 def cli(config):
     pass
 
-# Top Level Groups    
+@click.command(name='version', help='Print the current version of CumulusCI')
+def version():
+    click.echo(cumulusci.__version__)
+
+@click.command(name='shell', help='Drop into a python shell')
+@pass_config
+def shell(config):
+    code.interact(local=dict(globals(), **locals()))
+
+# Top Level Groups
 @click.group('project', help="Commands for interacting with project repository configurations")
 @pass_config
 def project(config):
@@ -105,13 +137,58 @@ cli.add_command(project)
 cli.add_command(org)
 cli.add_command(task)
 cli.add_command(flow)
-       
+cli.add_command(version)
+cli.add_command(shell)
+
 # Commands for group: project
 
-@click.command(name='init', help="Initialize a new project for use with the cumulusci toolbelt")
-@click.option('--name', help="The project's package name", prompt=True)
+@click.command(name='init',
+    help="Initialize a new project for use with the cumulusci toolbelt",
+)
+@click.option('--name',
+    help="The project's package name",
+    prompt=True,
+)
+@click.option('--package-name',
+    help="The project's package name",
+    prompt=True,
+)
+@click.option('--package-namespace',
+    help="The project's package namespace",
+    prompt=True,
+)
+@click.option('--package-api-version',
+    help="The Salesforce API verson for the package",
+    prompt=True,
+    default=CLI_CONFIG.global_config.project__package__api_version,
+)
+@click.option('--git-prefix-feature',
+    help="The branch prefix for all feature branches",
+    prompt=True,
+    default=CLI_CONFIG.global_config.project__git__prefix_feature,
+)
+@click.option('--git-default-branch',
+    help="The default branch in the repository",
+    prompt=True,
+    default=CLI_CONFIG.global_config.project__git__default_branch,
+)
+@click.option('--git-prefix-beta',
+    help="The tag prefix for beta release tags",
+    prompt=True,
+    default=CLI_CONFIG.global_config.project__git__prefix_beta,
+)
+@click.option('--git-prefix-release',
+    help="The tag prefix for production release tags",
+    prompt=True,
+    default=CLI_CONFIG.global_config.project__git__prefix_release,
+)
+@click.option('--test-namematch',
+    help="The SOQL format like query for selecting Apex tests.  % is wildcard",
+    prompt=True,
+    default=CLI_CONFIG.global_config.project__test__namematch,
+)
 @pass_config
-def project_init(config, name):
+def project_init(config, name, package_name, package_namespace, package_api_version, git_prefix_feature, git_default_branch, git_prefix_beta, git_prefix_release, test_namematch):
     if not os.path.isdir('.git'):
         click.echo("You are not in the root of a Git repository")
 
@@ -120,8 +197,48 @@ def project_init(config, name):
 
     f_yml = open('cumulusci.yml','w')
 
-    yml_config = 'project:\n    name: {}\n'.format(name)
-    f_yml.write(yml_config)
+    yml_config = []
+    # project:
+    yml_config.append('project:')
+    yml_config.append('    name: {}'.format(name))
+
+    #     package:
+    package_config = []
+    if package_name and package_name != config.global_config.project__package__name:
+        package_config.append('        name: {}'.format(package_name))
+    if package_namespace and package_namespace != config.global_config.project__package__namespace:
+        package_config.append('        namespace: {}'.format(package_namespace))
+    if package_api_version and package_api_version != config.global_config.project__package__api_version:
+        package_config.append('        api_version: {}'.format(package_api_version))
+    if package_config:
+        yml_config.append('    package:')
+        yml_config.extend(package_config)
+
+    #     git:
+    git_config = []
+    if git_prefix_feature and git_prefix_feature != config.global_config.project__git__prefix_feature:
+        git_config.append('        feature: {}'.format(git_prefix_feature))
+    if git_default_branch and git_default_branch != config.global_config.project__git__default_branch:
+        git_config.append('        branch: {}'.format(git_default_branch))
+    if git_prefix_beta and git_prefix_beta != config.global_config.project__git__prefix_beta:
+        git_config.append('        beta: {}'.format(git_prefix_beta))
+    if git_prefix_release and git_prefix_release != config.global_config.project__git__prefix_release:
+        git_config.append('        release: {}'.format(git_prefix_release))
+    if git_config:
+        yml_config.append('    git:')
+        yml_config.extend(git_config)
+
+
+    #     test:
+    test_config = []
+    if test_namematch and test_namematch != config.global_config.project__test__namematch:
+        test_config.append('        namematch: {}'.format(test_namematch))
+    if test_config:
+        yml_config.append('    test:')
+        yml_config.extend(test_config)
+
+    yml_config.append('')
+    f_yml.write('\n'.join(yml_config))
 
     click.echo("Your project is now initialized for use with CumulusCI")
     click.echo("You can use the project edit command to edit the project's config file")
@@ -145,7 +262,7 @@ def project_connect_github(config, username, password, email):
         'email': email,
     }))
     click.echo('Github is now configured for this project')
-        
+
 @click.command(name='show_github', help="Prints the current Github configuration for this project")
 @pass_config
 def project_show_github(config):
@@ -155,7 +272,7 @@ def project_show_github(config):
         click.echo(pretty_dict(github.config))
     except GithubNotConfigured:
         click.echo('Github is not configured for this project.  Use project connect_github to configure.')
-    
+
 
 @click.command(name='connect_mrbelvedere', help="Configure this project for mrbelvedere tasks")
 @click.option('--base_url', help="The base url for your mrbelvedere instance", prompt=True)
@@ -168,7 +285,7 @@ def project_connect_mrbelvedere(config, base_url, api_key):
         'api_key': api_key,
     }))
     click.echo('Mrbelvedere is now configured for this project')
-        
+
 @click.command(name='show_mrbelvedere', help="Prints the current mrbelvedere configuration for this project")
 @pass_config
 def project_show_mrbelvedere(config):
@@ -178,7 +295,7 @@ def project_show_mrbelvedere(config):
         click.echo(pretty_dict(mrbelvedere.config))
     except MrbelvedereNotConfigured:
         click.echo('mrbelvedere is not configured for this project.  Use project connect_mrbelvedere to configure.')
-    
+
 @click.command(name='connect_apextestsdb', help="Configure this project for ApexTestsDB tasks")
 @click.option('--base_url', help="The base url for your ApexTestsDB instance", prompt=True)
 @click.option('--user-id', help="The user id to use when connecting to ApexTestsDB.", prompt=True)
@@ -192,7 +309,7 @@ def project_connect_apextestsdb(config, base_url, user_id, token):
         'token': token,
     }))
     click.echo('ApexTestsDB is now configured for this project')
-        
+
 @click.command(name='show_apextestsdb', help="Prints the current ApexTestsDB configuration for this project")
 @pass_config
 def project_show_apextestsdb(config):
@@ -202,7 +319,7 @@ def project_show_apextestsdb(config):
         click.echo(pretty_dict(apextestsdb.config))
     except ApexTestsDBNotConfigured:
         click.echo('ApexTestsDB is not configured for this project.  Use project connect_apextestsdb to configure.')
-    
+
 
 @click.command(name='list', help="List projects and their locations")
 @pass_config
@@ -234,7 +351,7 @@ def org_browser(config, org_name):
 
     org_config = config.project_config.get_org(org_name)
     org_config.refresh_oauth_token(config.keychain.get_connected_app())
-    
+
     webbrowser.open(org_config.start_url)
 
 @click.command(name='connect', help="Connects a new org's credentials using OAuth Web Flow")
@@ -245,17 +362,17 @@ def org_connect(config, org_name, sandbox):
     check_connected_app(config)
 
     connected_app = config.keychain.get_connected_app()
-    
+
     oauth_capture = CaptureSalesforceOAuth(
         client_id = connected_app.client_id,
         client_secret = connected_app.client_secret,
         callback_url = connected_app.callback_url,
         sandbox = sandbox,
         scope = 'web full refresh_token'
-    ) 
+    )
     oauth_dict = oauth_capture()
     org_config = OrgConfig(oauth_dict)
-    
+
     config.keychain.set_org(org_name, org_config)
 
 @click.command(name='default', help="Sets an org as the default org for tasks and flows")
@@ -284,12 +401,16 @@ def org_info(config, org_name):
 @pass_config
 def org_list(config):
     check_connected_app(config)
+    data = []
+    headers = ['org','is_default']
     for org in config.project_config.list_orgs():
         org_config = config.project_config.get_org(org)
         if org_config.default:
-            click.echo('  * {}'.format(org))
+            data.append((org, '*'))
         else:
-            click.echo('    {}'.format(org))
+            data.append((org, ''))
+    table = Table(data, headers)
+    click.echo(table)
 
 @click.command(name='connected_app', help="Displays the ConnectedApp info used for OAuth connections")
 @pass_config
@@ -326,8 +447,12 @@ org.add_command(org_config_connected_app)
 @pass_config
 def task_list(config):
     check_project_config(config)
+    data = []
+    headers = ['task', 'description']
     for task in config.project_config.list_tasks():
-        click.echo('{name}: {description}'.format(**task))
+        data.append((task['name'], task['description']))
+    table = Table(data, headers)
+    click.echo(table)
 
 @click.command(name='info', help="Displays information for a task")
 @click.argument('task_name')
@@ -337,7 +462,7 @@ def task_info(config, task_name):
     task_config = getattr(config.project_config, 'tasks__{}'.format(task_name))
     class_path = task_config.get('class_path')
     task_class = import_class(class_path)
-    
+
     # General task info
     click.echo('Description: {}'.format(task_config.get('description')))
     click.echo('Class: {}'.format(task_config.get('class_path')))
@@ -354,13 +479,15 @@ def task_info(config, task_name):
     task_options = getattr(task_class, 'task_options', {})
     if task_options:
         click.echo('')
-        click.echo('Options:')
+        data = []
+        headers = ['Option', 'Required', 'Description']
         for key, option in task_options.items():
             if option.get('required'):
-                click.echo('  * {}: {}'.format(key, option.get('description')))
-        for key, option in getattr(task_class, 'task_options', {}).items():
-            if not option.get('required'):
-                click.echo('    {}: {}'.format(key, option.get('description')))
+                data.append((key, '*', option.get('description')))
+            else:
+                data.append((key, '', option.get('description')))
+        table = Table(data, headers)
+        click.echo(table)
 
 @click.command(name='run', help="Runs a task")
 @click.argument('task_name')
@@ -403,14 +530,19 @@ def task_run(config, task_name, org, o):
             task_config['options'][name] = value
 
     task_config = TaskConfig(task_config)
-    
-    # Create and run the task    
+
+    # Create and run the task
     try:
         task = task_class(config.project_config, task_config, org_config = org_config)
     except TaskRequiresSalesforceOrg as e:
         raise click.UsageError('This task requires a salesforce org.  Use org default <name> to set a default org or pass the org name with the --org option')
-        
-    task()
+    except TaskOptionsError as e:
+        raise click.UsageError(e.message)
+
+    try:
+        task()
+    except TaskOptionsError as e:
+        raise click.UsageError(e.message)
 
 # Add the task commands to the task group
 task.add_command(task_list)
@@ -422,7 +554,13 @@ task.add_command(task_run)
 @pass_config
 def flow_list(config):
     check_project_config(config)
-    click.echo(config.project_config.flows)
+    data = []
+    headers = ['flow', 'description']
+    for flow in config.project_config.flows:
+        description = getattr(config.project_config, 'flows__{}__description'.format(flow))
+        data.append((flow, description))
+    table = Table(data, headers)
+    click.echo(table)
 
 @click.command(name='info', help="Displays information for a flow")
 @click.argument('flow_name')
@@ -450,9 +588,19 @@ def flow_run(config, flow_name, org):
     class_path = flow_config.get('class_path', 'cumulusci.core.flows.BaseFlow')
     flow_class = import_class(class_path)
 
-    # Create and run the flow 
-    flow = flow_class(config.project_config, flow_config, org_config)
-    flow()
+    # Create the flow and handle initialization exceptions
+    try:
+        flow = flow_class(config.project_config, flow_config, org_config)
+    except TaskRequiresSalesforceOrg as e:
+        raise click.UsageError('This flow requires a salesforce org.  Use org default <name> to set a default org or pass the org name with the --org option')
+    except TaskOptionsError as e:
+        raise click.UsageError(e.message)
+
+    # Run the flow and handle exceptions
+    try:
+        flow()
+    except TaskOptionsError as e:
+        raise click.UsageError(e.message)
 
 flow.add_command(flow_list)
 flow.add_command(flow_info)